+++ src/main/java/de/maxhenkel/reap/Events.java
@@ -1,11 +1,11 @@
 package de.maxhenkel.reap;
 
 import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.util.ActionResultType;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.IWorld;
 import net.minecraftforge.event.entity.player.PlayerInteractEvent;
 import net.minecraftforge.event.world.BlockEvent;
-import net.minecraftforge.eventbus.api.Event;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 
@@ -29,10 +29,9 @@ public class Events {
         boolean success = Harvester.harvest(clickedBlock, player);
 
         if (success && event.isCancelable()) {
-            event.setResult(Event.Result.DENY);
+            event.setCancellationResult(ActionResultType.SUCCESS);
             event.setCanceled(true);
         }
-
     }
 
     @SubscribeEvent