+++ src/main/java/de/maxhenkel/reap/Events.java
@@ -2,8 +2,7 @@ package de.maxhenkel.reap;
 
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.util.ActionResultType;
-import net.minecraft.util.math.BlockPos;
-import net.minecraft.world.IWorld;
+import net.minecraft.world.World;
 import net.minecraftforge.event.entity.player.PlayerInteractEvent;
 import net.minecraftforge.event.world.BlockEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -14,21 +13,7 @@ public class Events {
 
     @SubscribeEvent
     public void onPlayerUse(PlayerInteractEvent.RightClickBlock event) {
-        if (event.isCanceled()) {
-            return;
-        }
-
-        PlayerEntity player = event.getPlayer();
-
-        if (player == null) {
-            return;
-        }
-
-        BlockPos clickedBlock = event.getPos();
-
-        boolean success = Harvester.harvest(clickedBlock, player);
-
-        if (success && event.isCancelable()) {
+        if (Harvester.harvest(event.getPos(), event.getPlayer())) {
             event.setCancellationResult(ActionResultType.SUCCESS);
             event.setCanceled(true);
         }
@@ -36,17 +21,10 @@ public class Events {
 
     @SubscribeEvent
     public void onBlockBreak(BlockEvent.BreakEvent event) {
-        IWorld world = event.getWorld();
         PlayerEntity player = event.getPlayer();
-
-        if (event.getWorld().isRemote() || player == null || event.isCanceled() || player.abilities.isCreativeMode) {
+        if (event.getWorld().isRemote() || player.abilities.isCreativeMode) {
             return;
         }
-
-        BlockPos pos = event.getPos();
-
-        TreeHarvester harvester = new TreeHarvester(pos, player, world);
-
-        harvester.harvest();
+        TreeHarvester.harvest(event.getPos(), player, (World) event.getWorld());
     }
 }