+++ changelog.md
@@ -1,1 +1,1 @@
-- Added dynamic tree breaking speed
+- Updated to 1.16.1